refactor: move all enums to enump.py
