+++ google/cloud/sql/connector/__init__.py
@@ -16,8 +16,8 @@ limitations under the License.
 
 from google.cloud.sql.connector.connector import Connector
 from google.cloud.sql.connector.connector import create_async_connector
-from google.cloud.sql.connector.instance import IPTypes
-from google.cloud.sql.connector.instance import RefreshStrategy
+from google.cloud.sql.connector.enums import IPTypes
+from google.cloud.sql.connector.enums import RefreshStrategy
 from google.cloud.sql.connector.version import __version__
 
 __all__ = [
